# server/docs/swagger.yaml
definitions:
...
  response.Email:
    properties:
      body:
        description: email body
        type: string
      subject:
        description: email subject
        type: string
      to:
        description: email recipient
        type: string
    type: object
...
  response.PageResult:
    properties:
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  system.SysParams:
    properties:
      ID:
        description: primary key ID
        type: integer
      createdAt:
        description: created at
        type: string
      desc:
        description: parameter description
        type: string
      key:
        description: parameter key
        type: string
      name:
        description: Parametername
        type: string
      updatedAt:
        description: updated at
        type: string
      value:
        description: parameter value
        type: string
    required:
    - key
    - name
    - value
    type: object
...
info:
  contact: {}
  description: Full-stack rapid development with Gin + Vue
  title: Gin-Vue-Admin Swagger API
  version: v2.9.1
paths:
...
  /email/emailTest:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"Sent successfully"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: send test email
      tags:
      - System
  /email/sendEmail:
    post:
      parameters:
      - description: send emailrequired parameters
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/response.Email'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"Sent successfully"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: send email
      tags:
      - System
...
  /sysParams/createSysParams:
    post:
      consumes:
      - application/json
      parameters:
      - description: Create parameter
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.SysParams'
      produces:
      - application/json
      responses:
        "200":
          description: Created successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Create parameter
      tags:
      - SysParams
  /sysParams/deleteSysParams:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Delete parameter
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.SysParams'
      produces:
      - application/json
      responses:
        "200":
          description: Deleted successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Delete parameter
      tags:
      - SysParams
  /sysParams/deleteSysParamsByIds:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: Batch deleted successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Batch delete parameters
      tags:
      - SysParams
  /sysParams/findSysParams:
    get:
      consumes:
      - application/json
      parameters:
      - description: primary key ID
        in: query
        name: ID
        type: integer
      - description: created at
        in: query
        name: createdAt
        type: string
      - description: parameter description
        in: query
        name: desc
        type: string
      - description: parameter key
        in: query
        name: key
        required: true
        type: string
      - description: Parametername
        in: query
        name: name
        required: true
        type: string
      - description: updated at
        in: query
        name: updatedAt
        type: string
      - description: parameter value
        in: query
        name: value
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Query successful
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/system.SysParams'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Find parameter by ID
      tags:
      - SysParams
  /sysParams/getSysParam:
    get:
      consumes:
      - application/json
      parameters:
      - description: key
        in: query
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/system.SysParams'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get parameter value by key
      tags:
      - SysParams
  /sysParams/getSysParamsList:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: endCreatedAt
        type: string
      - in: query
        name: key
        type: string
      - description: keyword
        in: query
        name: keyword
        type: string
      - in: query
        name: name
        type: string
      - description: page number
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: pageSize
        type: integer
      - in: query
        name: startCreatedAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get parameter list with pagination
      tags:
      - SysParams
  /sysParams/updateSysParams:
    put:
      consumes:
      - application/json
      parameters:
      - description: Update parameter
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.SysParams'
      produces:
      - application/json
      responses:
        "200":
          description: Updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Update parameter
      tags:
      - SysParams
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"
...